# (perf) optimising stripping bearer token