Should speed up JWT & Auth Token a little bit.

```
benchcmp old.txt new.txt
benchmark                  old ns/op     new ns/op     delta
BenchmarkStripBearer-4     165           38.4          -76.73%

benchmark                  old allocs     new allocs     delta
BenchmarkStripBearer-4     2              1              -50.00%

benchmark                  old bytes     new bytes     delta
BenchmarkStripBearer-4     96            8             -91.67%
```